Fix: fix minor bugs in the application.

+++ src/main/java/uz/pdp/lionestatebot/models/entity/SavedProperty.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
+import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "saved_properties")
@@ -15,7 +15,7 @@ public class SavedProperty {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     private BotUser user;
